codigo_backup_20260226_130308/backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
gunicorn>=21.0
